% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend-imputed-dataframe.R
\name{extend_imputed_dataframe}
\alias{extend_imputed_dataframe}
\title{Extension of time series in imputed dataframe}
\usage{
extend_imputed_dataframe(idf, length_in_months = 25)
}
\arguments{
\item{idf}{Imputed dataframe.}

\item{length_in_months}{Minimum number of months of the extended output.}
}
\value{
Imputed dataframe, i.e. a cooked dataframe with a 3rd column indicating if each sample has been imputed or not.
}
\description{
Replicate the existing values of a periodic time series to obtain a time series with the desired length in months.
}
