% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-features-of-datasets-in-folder.R
\name{get_features_of_datasets_in_folder}
\alias{get_features_of_datasets_in_folder}
\title{Features of datasets in a folder}
\usage{
get_features_of_datasets_in_folder(
  folder_path,
  from_date,
  to_date,
  dset_key,
  allowed_na,
  type_of_analysis,
  output_folder_path = NULL
)
}
\arguments{
\item{folder_path}{String with the absolute path to the dataset folder (ending in `/`).}

\item{from_date}{Initial date and time of the interval. Either a `POSIXct` class in the GMT time zone OR the string `first`.}

\item{to_date}{Final date and time of the interval. Either a `POSIXct` class in the GMT time zone OR the string `last`.}

\item{dset_key}{Key of the dataset.}

\item{allowed_na}{A numerical value between 0 and 1. It represents the maximum percentage of admissible `NA` values in the cooked dataframe for which the feature extraction is performed. The `NA` values will be STL-interpolated (using ``forecast::na.interp``) prior to the feature extraction.}

\item{type_of_analysis}{A string indicating the type of analysis: either `basic` or `extra`.}

\item{output_folder_path}{String with the absolute path to the output folder (ending in `/`).}
}
\value{
List of dataframes of (1) extracted features and (2) accepted and (3) rejected files for feature extraction.
}
\description{
Get features of all datasets contained in a folder.
}
