% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-load-and-processing.R
\name{impute_cooked_dataframe}
\alias{impute_cooked_dataframe}
\title{Imputation of missing samples in cooked dataframe}
\usage{
impute_cooked_dataframe(
  cdf,
  season,
  short_gap,
  short_algorithm = "interpolation",
  long_algorithm = "locf"
)
}
\arguments{
\item{cdf}{Cooked dataframe.}

\item{season}{Seasonal period (e.g. 1 week) in number of samples.}

\item{short_gap}{Number of samples considered as a short gap.}

\item{short_algorithm}{Algorithm used to impute short gaps.}

\item{long_algorithm}{Algorithm used to impute long gaps.}
}
\value{
Imputed dataframe, i.e. a cooked dataframe with a 3rd column indicating if each sample has been imputed or not.
}
\description{
Impute missing samples in a cooked dataframe. It can use an algorithm for short gaps (e.g. "interpolation") and another one for longer gaps (e.g. "locf").
}
