% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning-analysis.R
\name{pca_from_features}
\alias{pca_from_features}
\title{PCA of features}
\usage{
pca_from_features(feats_folder, otp = NULL, ftp)
}
\arguments{
\item{feats_folder}{String with the path to the folder of the features file.}

\item{otp}{Vector of observations to plot. `NULL` indicates all observations.}

\item{ftp}{Vector of features to plot.}
}
\value{
A list with class `princomp`.
}
\description{
Compute and plot PCA from a dataframe of features.
}
