% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{quantiles}
\alias{quantiles}
\title{Features of quantiles}
\usage{
quantiles(x)
}
\arguments{
\item{x}{Time series of class `msts`.}
}
\value{
A list with the minimum, lower quartile, median, upper quartile and maximum of the time series.
}
\description{
Compute the minimum, lower quartile, median, upper quartile and maximum of a time series.
}
