% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-tools.R
\name{plot_pca}
\alias{plot_pca}
\title{Plotting of PCA scores}
\usage{
plot_pca(
  pca_sc,
  feats_folder,
  axis_x,
  axis_y,
  color_by_SE_vars = FALSE,
  SE_data_file = NULL,
  get_point_identity = FALSE
)
}
\arguments{
\item{pca_sc}{A dataframe with PCA scores.}

\item{feats_folder}{String with the path to the folder of the features file.}

\item{axis_x}{Integer indicating the principal component to be plotted as axis x.}

\item{axis_y}{Integer indicating the principal component to be plotted as axis y.}

\item{color_by_SE_vars}{Boolean indicating if plotted points must be colored according to the socioeconomical variables.}

\item{SE_data_file}{String with the path to the file with the socioeconomical variables.}
}
\value{
Plotting of the PCA scores.
}
\description{
Compute and plot PCA from a dataframe of features.
}
